requirements.txt
# === Core Dependencies ===
# MCP Server
fastmcp>=0.1.0

# Browser Automation (вместо API)
playwright>=1.40.0

# AI/ML Dependencies
# TF-IDF Scoring
scikit-learn>=1.3.0
numpy>=1.24.0

# Letter Generation (OpenAI API)
openai>=1.10.0

# === Analytics ===
pandas>=2.1.0
matplotlib>=3.8.0

# === Testing ===
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-playwright>=0.4.3

# === Code Quality ===
ruff>=0.1.0
mypy>=1.7.0

# === Utilities ===
click>=8.1.0  # CLI interface (optional)
rich>=13.7.0  # Console formatting
tenacity>=8.2.0  # Retry logic
python-dotenv>=1.0.0  # Environment variables
pydantic>=2.5.0  # Data validation
pydantic-settings>=2.1.0  # Settings management
